requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
sentence-transformers
python-dotenv
requests
openai

langchain_community
langchain_core
langchain_text_splitters